Optimize VarHandle.getAndAdd calls for longs. Must pass in a long instead of an int, otherwise a slow path is taken to perform the conversion.

+++ src/main/java/org/cojen/tupl/core/TransactionContext.java
@@ -56,7 +56,7 @@ final class TransactionContext extends Latch implements Flushable {
         }
     }
 
-    private final int mTxnStride;
+    private final long mTxnStride;
 
     // Access to these fields is protected by synchronizing on this context object.
     private long mInitialTxnId;
